requirements.txt
asgiref==3.2.10
Django==3.0.9
sqlparse==0.2.4
django-tinymce
Pillow==7.0.0

